+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphRenderer.tsx
@@ -290,7 +290,6 @@ class FlameGraphRenderer extends React.Component<
 
   updateFlamebearerData() {
     const flamebearer = mountFlamebearer(this.props);
-
     this.setState({ flamebearer });
   }
 
@@ -338,6 +337,8 @@ class FlameGraphRenderer extends React.Component<
     //      this.props.viewSide
     //    );
 
+    const toolbarVisible = this.shouldShowToolbar();
+
     const flameGraphPane =
       this.state.flamebearer && dataExists ? (
         <Graph
@@ -354,6 +355,7 @@ class FlameGraphRenderer extends React.Component<
           onReset={this.onReset}
           isDirty={this.isDirty}
           palette={this.state.palette}
+          toolbarVisible={toolbarVisible}
           setPalette={(p) =>
             this.setState({
               palette: p,
@@ -367,7 +369,7 @@ class FlameGraphRenderer extends React.Component<
     return (
       <div>
         <div>
-          {this.shouldShowToolbar() && (
+          {toolbarVisible && (
             <Toolbar
               renderLogo={this.props.renderLogo || false}
               disableChangingDisplay={!!this.props.onlyDisplay}